Fixed bug in PiecewiseConstant discovered in EONAV
Clipping was not done properly: the cast to int must be made after
taking the min otherwise we risk an overflow for huge doubles.

git-svn-id: http://130.66.124.6/svn/maroff/Product/Cpp/Tools/interpolation@1628 300066c3-8af3-463d-8cc0-c0233cc64ac1

+++ interpolation/src/PiecewiseConstant.cpp
@@ -20,7 +20,7 @@ Interpolator(xmin_,xmax_,y_)
 
 void PiecewiseConstant::update_coefficients_if_necessary(const double val)
 {
-    idx = std::max(0,std::min((int)floor((val-xmin)/(xmax-xmin)*n),(int)n-1));
+    idx = std::max(0,(int)std::min(floor((val-xmin)/(xmax-xmin)*n),(double)n-1));
 }
 
 void PiecewiseConstant::compute_coefficients_for_ith_interval(const double x0, const size_t i)
